refactor packages add unit tests
